(chore/refactor:NewTickets) move button click handlers to useNewTickets
